refactor : 당첨번호+보너스볼 질문은 InputView로 이동
