+++ README.md
@@ -103,4 +103,5 @@
   - [x] 리팩토링 및 테스트
     - 결과값 계산 확인
    
+  - [x] 당첨번호+보너스볼 질문은 InputView로
