+++ src/main/java/lotto/view/ResultView.java
@@ -13,8 +13,6 @@ import java.util.Map;
  */
 public class ResultView {
 
-    private static final String QUESTION_WINNING_NUMBER = "지난 주 당첨 번호를 입력해 주세요.";
-    private static final String QUESTION_BONUS_BALL = "보너스 볼을 입력해 주세요.";
     private static final String MESSAGE_STATISTICS = "당첨 통계";
     private static final String RESULT_STATISTICS = "%d개 일치 (%d원)";
     private static final String RESULT_NUMBER = "- %d개";
@@ -25,23 +23,6 @@ public class ResultView {
         lottos.list().forEach(lotto -> System.out.println(lotto.number()));
         System.out.println();
     }
-
-    // TODO inputView 로 옮겨야함.
-    public static WinningNumber askWinningNumberAndBonus() {
-        return WinningNumber.of(askWinningNumber(), askBonusBall());
-    }
-
-    private static String askWinningNumber() {
-        System.out.println(QUESTION_WINNING_NUMBER);
-        return LottoGameController.SCANNER.nextLine();
-    }
-
-    private static String askBonusBall() {
-        System.out.println(QUESTION_BONUS_BALL);
-        return LottoGameController.SCANNER.nextLine();
-    }
-
-
 
 
     public static void showResults(WinningStatistics winningStatistics, int amount) {
